refactor: 프로필 이미지 리사이징 및 적용
