+++ frontend/src/domains/user/home/constants/FeedbackForm.ts
@@ -1,11 +1,11 @@
-import avatar1 from '@/assets/images/avatar1.png';
-import avatar2 from '@/assets/images/avatar2.png';
-import avatar3 from '@/assets/images/avatar3.png';
-import avatar4 from '@/assets/images/avatar4.png';
-import avatar5 from '@/assets/images/avatar5.png';
-import avatar6 from '@/assets/images/avatar6.png';
-import avatar7 from '@/assets/images/avatar7.png';
-import avatar8 from '@/assets/images/avatar8.png';
+import avatar1 from '@/assets/images/avatar1-64x64.png';
+import avatar2 from '@/assets/images/avatar2-64x64.png';
+import avatar3 from '@/assets/images/avatar3-64x64.png';
+import avatar4 from '@/assets/images/avatar4-64x64.png';
+import avatar5 from '@/assets/images/avatar5-64x64.png';
+import avatar6 from '@/assets/images/avatar6-64x64.png';
+import avatar7 from '@/assets/images/avatar7-64x64.png';
+import avatar8 from '@/assets/images/avatar8-64x64.png';
 
 export const FEEDBACK_FORM_CONSTANTS = {
   ADJECTIVES: [
