+++ frontend/src/utils/getUserImages.ts
@@ -1,10 +1,10 @@
-import IMG1 from '@/assets/images/avatar1.png';
-import IMG2 from '@/assets/images/avatar2.png';
-import IMG3 from '@/assets/images/avatar3.png';
-import IMG4 from '@/assets/images/avatar4.png';
-import IMG5 from '@/assets/images/avatar5.png';
-import IMG6 from '@/assets/images/avatar6.png';
-import IMG7 from '@/assets/images/avatar7.png';
-import IMG8 from '@/assets/images/avatar8.png';
+import IMG1 from '@/assets/images/avatar1-64x64.png';
+import IMG2 from '@/assets/images/avatar2-64x64.png';
+import IMG3 from '@/assets/images/avatar3-64x64.png';
+import IMG4 from '@/assets/images/avatar4-64x64.png';
+import IMG5 from '@/assets/images/avatar5-64x64.png';
+import IMG6 from '@/assets/images/avatar6-64x64.png';
+import IMG7 from '@/assets/images/avatar7-64x64.png';
+import IMG8 from '@/assets/images/avatar8-64x64.png';
 
 export const userImages = [IMG1, IMG2, IMG3, IMG4, IMG5, IMG6, IMG7, IMG8];
